workloads/thumbnailer-python/pyproject.toml
[project]
name = "function"
version = "0.1.0"
description = "Knative function for resizing images with MinIO integration"
authors = [
  { name = "Your Name", email = "your.email@example.com" }
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "minio>=7.1.3",    # For MinIO integration
  "Pillow>=9.0.0"    # For image processing
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",      # For testing
  "pytest-asyncio>=0.21"  # For async testing
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
